refactor(ai): extract blobToBase64 to shared utils/fileEncoding
Deduplicates the identical helper between GuidedLearningAIGenerator and
DriveImagePicker into a single `blobToBase64` export. Addresses Gemini
code review feedback on #1368.
